Replace active result sources with active counters
This is how I originally intended to implement this.
I realised that keeping sources around is often
insufficient for determining whether said sources
are active, which was a mistake on my part.

Instead the client now counts up when a source is
started (subscribed to) and counts down when a source
ends. This should make the teardown accurate,
especially when a synchronous End is sent after
Start before onStart triggers or when a source is
kept around by a user which can often be the case.