Implement reading & traversing resolver results and pagination helpers (#91)
* Add parentTypeName and fieldName props to ResolveInfo

* Add initial Relay pagination helper

* Implement copy-first resolver result reading

This replaces the simple resolver result reader with
a "drilling" result reader, `readResolverResult` and
`resolveResolverResult`.

The former is similar to `readSelection` and the
latter is similar to `resolveLink`.

Instead of either resolving an entity or copying
the result of the resolver, we now read the data
from the resolver result if no key exists in the
store.

* Add tests for new nested, unkeyed resolvers

* Update relayPagination helper to use resolve properly

* Add more selection logic to readResolverResult

* Remove links key fallback from resolveResolverResult

This is actually incorrect. If the user returns an entity somewhere
in their resolver result, there's no reason to believe that it's the
same as the one the cache is pointing at (quite the opposite actually)

Instead we should still rely on entity keying.

* Add new test nesting a keyable entity inside an unkeyable one

* Fix relayPagination quirks and adapt to changes

* Add tests for relayPagination helper

* Add bundling for extras folder

This doesn't yet add any entrypoint aliasing for automatic code
splitting to Rollup since that's not worth it and not confirmed
to work well yet.

* Remove leftover comment

* Add dependency tracking to readResolverResult

* Add queryDependencies to updates in updateCacheWithResult

This is a significant change, where we may update more
queries than we have before. It adds queryDependencies
to the dependencies that are reexecuted when the cache is
updated, apart from just writeDependencies.

This is necessary because cache resolvers can alter
the data that is returned tremendously.

It should still be safe to do this though, because for
queries without any cache resolvers the queried data
should match the written data 1:1.

For data with cache resolvers that is queried however,
we will get some extra dependencies, which do need to
update.

* Fix hasPreviousPage flag usage

* Add connection tracking

This introduces a Pessimism.Map that tracks each
variables that have been stored for a given entity
field, while also pointing at the link where that
field's data is being stored.

* Add resolveConnections method to Store

* WIP: Start new relayPagination implementation

* (feat) - find the connection we are on

* Implement Relay algorithm for concatenating pages

See Relay source for original algorithm:
https://github.com/facebook/relay/blob/7153ce1/packages/relay-runtime/store/RelayConnectionResolver.js#L66-L102

* Deduplicate edge nodes as they're being concatenated

* Add failing test for fetching duplicate edges

* Add failing test for simultaneous forwards- and backwards pagination

* Fix up backwards/forwards pagination test

* Add inward/outward merging and tests

* Simplify bail-out case

This removes the need for our key helpers

* Add parentFieldKey to ResolveInfo

This eliminates the need for having joinKeys and keyOfField in userland

* Add abstract Cache interface