(request-policy) - Fix requestPolicyExchange usage with persisted caching (#1742)
* Ignore files generated by JetBrains IDEs like Intellij

* (request-policy) - Fix TTL being set on cache hits

Do not set the TTL unless cache outcome is "miss". If the first result
to come through the requestPolicyExchange was a cache-hit then the ttl was incorrectly
set initially based on this cache hit and no network request was made. This happens when using
persisted caching with the graphcache exchange since the very first result can be a cache hit.

https://github.com/FormidableLabs/urql/issues/1644

* Add changeset